requirements.txt
# Core RAG dependencies
FlagEmbedding==1.3.3
qdrant-client==1.12.1
langchain==0.3.14
langchain-community==0.3.14

# Document processing
pdfplumber==0.11.4
python-docx==1.1.2
unstructured==0.16.12

# LLM (local via Ollama)
ollama==0.4.4

# API layer
flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0

# Evaluation
ragas==0.2.6
datasets==3.2.0

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
loguru==0.7.3
numpy==1.26.4
arxiv==2.1.0
